ADA 2 Daniel Rosales
ADA2 Daniel Rosales

+++ app/Http/Controllers/ProductController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Product;
 use Illuminate\Http\Request;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class ProductController extends Controller
 {
@@ -34,13 +35,37 @@ class ProductController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {
-        $product = Product::create($request->all());
+     {
+        if (is_null($request->name)) {
+                return response()->json([
+                    "errors"=> ["code"=> "ERROR-2",
+                    "title"=>  "Unprocessable Entity",
+                    ]]  , 422);
+        }elseif (is_null($request->price)) {
+                return response()->json([
+                    "errors"=> ["code"=> "ERROR-3",
+                    "title"=>  "Unprocessable Entity",
+                    ]]  , 422);
+        }elseif (!(is_numeric($request->price))) {
+                return response()->json([
+                   "errors"=> ["code"=> "ERROR-4",
+                   "title"=>  "Unprocessable Entity",
+                   ]]  , 422);
+        }elseif (($request->price)<=0) {
+                return response()->json([
+                    "errors"=> ["code"=> "ERROR-5",
+                    "title"=>  "Unprocessable Entity",
+                    ]]  , 422);
+        }else {
+            // Create a new product
+            $product = Product::create($request->all());
 
-        // Return a response with a product json
-        // representation and a 201 status code   
-        return response()->json($product,201);
-    }
+            // Return a response with a product json
+            // representation and a 201 status code
+            return response()->json($product,201);
+        }
+
+     }
 
     /**
      * Display the specified resource.
@@ -63,9 +88,18 @@ class ProductController extends Controller
      */
     public function showItem(Product $product, $id)
     {
-        // Return a response with a product json
-        // representation and a 200 status code   
-        return $product->findorfail($id);
+        try {
+            $producto = Product::findOrFail($id);
+            return $producto;
+        } catch (ModelNotFoundException $exception) {
+            $type = explode("\\", get_class($exception));
+            $type = $type[count($type)-1];
+            return response()->json([
+                "errors"=> ["code"=> "ERROR-2",
+                "title"=>  "Not Found",
+                "type" => $type
+                ]]  , 404);
+        }
     }
 
     /**
@@ -88,13 +122,25 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $product = Product::findorfail($id);
-        $product->name=$request->name;
-        $product->price=$request->price;
-        $product->save();
-        // Return a response with a product json
-        // representation and a 201 status code   
-        return response()->json($product,200);
+        if (!(is_numeric($request->price))) {
+                return response()->json([
+                "errors"=> ["code"=> "ERROR-1",
+                "title"=>  "Unprocessable Entity",
+                ]]  , 422);
+        }elseif (($request->price)<=0) {
+                return response()->json([
+                    "errors"=> ["code"=> "ERROR-5",
+                    "title"=>  "Unprocessable Entity",
+                    ]]  , 422);
+        }else{
+            $product = Product::findorfail($id);
+            $product->name=$request->name;
+            $product->price=$request->price;
+            $product->save();
+            // Return a response with a product json
+            // representation and a 201 status code   
+            return response()->json($product,200);
+        }
     }
 
     /**
@@ -107,6 +153,15 @@ class ProductController extends Controller
     {
         // Return a response with a product json
         // representation and a 200 status code   
-        return $product->destroy($id);
+        if(!find($id)){
+            return response()->json([
+                "errors"=> ["code"=> "ERROR-2",
+                "title"=>  "Not Found",
+                ]]  , 404);
+        }
+        else{
+            Product::findOrFail($id);
+            return $product->destroy($id);
+        }
     }
 }
